add components and split sections
