+++ components/Slider/SmallSteps.js
@@ -1,0 +1,24 @@
+import * as React from "react";
+import Box from "@mui/material/Box";
+import Slider from "@mui/material/Slider";
+
+function valuetext(value) {
+  return `${value}°C`;
+}
+
+export default function DiscreteSliderSteps() {
+  return (
+    <Box sx={{ width: 300 }}>
+      <Slider
+        aria-label="Small steps"
+        defaultValue={0.00000005}
+        getAriaValueText={valuetext}
+        step={0.00000001}
+        marks
+        min={-0.00000005}
+        max={0.0000001}
+        valueLabelDisplay="auto"
+      />
+    </Box>
+  );
+}
