+++ package.json
@@ -22,6 +22,10 @@
     "@ckeditor/ckeditor5-paragraph": "^34.0.0",
     "@ckeditor/ckeditor5-react": "^5.0.1",
     "@ckeditor/ckeditor5-theme-lark": "^34.0.0",
+    "@date-io/date-fns": "^2.14.0",
+    "@date-io/dayjs": "^2.14.0",
+    "@date-io/luxon": "^2.14.0",
+    "@date-io/moment": "^2.14.0",
     "@elastic/elasticsearch": "^8.1.0",
     "@emotion/react": "^11.9.0",
     "@emotion/styled": "^11.8.1",
@@ -30,6 +34,7 @@
     "@mui/icons-material": "^5.6.2",
     "@mui/material": "^5.6.4",
     "@mui/styles": "^5.6.2",
+    "@mui/x-date-pickers": "^5.0.0-alpha.4",
     "@next-auth/prisma-adapter": "^1.0.3",
     "@next/eslint-plugin-next": "^12.1.6",
     "@prisma/client": "^3.13.0",
@@ -39,6 +44,7 @@
     "config": "^3.3.7",
     "csv-loader": "^3.0.3",
     "d3": "^7.4.4",
+    "date-fns": "^2.28.0",
     "dotenv": "^16.0.0",
     "markdown-to-jsx": "^7.1.7",
     "material-ui-popup-state": "^3.1.1",
